+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [3.2.0] - 2020-05-10 
+### Added
+- Only show indicator when Read-Only (issue [#15](https://github.com/alefragnani/vscode-read-only-indicator/issues/15)) - (Thanks to @chrisant996 [PR #24](https://github.com/alefragnani/vscode-read-only-indicator/pull/24))
+- Settings to choose the colors of the Status Bar text, using `workbench.colorCustomizations` (Thanks to @chrisant996 [PR #24](https://github.com/alefragnani/vscode-read-only-indicator/pull/24))
+
 ## [3.1.0] - 2019-03-01 
 ### Added
 - Settings changes detection (issue [#21](https://github.com/alefragnani/vscode-read-only-indicator/issues/21))